+++ markdown/extensions/admonition.py
@@ -41,7 +41,15 @@ class AdmonitionExtension(Extension):
     """ Admonition extension for Python-Markdown. """
 
     def extendMarkdown(self, md):
-        """ Add Admonition to Markdown instance. """
+        """
+        Register the processor.
+
+        | Class Instance                                                              | Registry                                                         | Name   | Priority |
+        | --------------------------------------------------------------------------- | ---------------------------------------------------------------- | ------ | :------: |
+        | [`AdmonitionProcessor`][markdown.extensions.admonition.AdmonitionProcessor] | [`blockprocessors`][markdown.blockprocessors.build_block_parser] | `admonition` | `105` |
+
+        """
+        # flake8: noqa: E501 47-49
         md.registerExtension(self)
 
         md.parser.blockprocessors.register(AdmonitionProcessor(md.parser), 'admonition', 105)