+++ markdown/extensions/wikilinks.py
@@ -48,6 +48,14 @@ class WikiLinkExtension(Extension):
         super().__init__(**kwargs)
 
     def extendMarkdown(self, md):
+        """ Register the processor.
+
+        | Class Instance                                                              | Registry                                                         | Name   | Priority |
+        | --------------------------------------------------------------------------- | ---------------------------------------------------------------- | ------ | :------: |
+        | [`WikiLinksInlineProcessor`][markdown.extensions.wikilinks.WikiLinksInlineProcessor] | [`inlinepatterns`][markdown.inlinepatterns.build_inlinepatterns] | `wikilink` | `75` |
+
+        """
+        # flake8: noqa: E501 53-55
         self.md = md
 
         # append to end of inline patterns